feat: Doxygen comments for UnitTests
